Create Lagrange Code + Rational Structs
(should test for accuracy, but jupyter didn't flag anything)
This code has structs for polynomials and rational functions. It also has a squarematrix struct which can be used for both square matrices and bivariate polynomials. The idea behind Lagrange is if I have some symmetric polynomial in T1, T2 with degree at most n in T1 and in T2, then if I have an (n+1)x(n+1) lattice of input points, I know a polynomial that vanishes at all of these points but one. I can then use that in a similar way to single variable Lagrange interpolation to get a polynomial in two variables based on some points.