Big changes to make inverse algo work
Reworked the new polynomial matrix struct - now, its the same memory layout as the old one, but the elements themselves are pointers to polynomials. This turns out to work nicely in terms of how well different parts of the algorithm plug into each other. The recursion is finished except for matrix multiplication which needs to be inserted into the code